Move node 4.1.7 and 4.1.8 tests to script to avoid errors
On running the scan for CIS 1.6, the scan got stuck in Error state with the following error:
```
level=fatal msg=\"error summarizing: \\nfailed to check if command: \\\"CAFILE=$(ps\\\" is valid exit status 2
```
It looks like the command isn't getting parsed. Moving the audit commands to scripts and running the
scripts seems to resolve this issue.