Changes for RKE hardened profile
All tests that aren't applicable to RKE are "skipped" with the reason in remediation. This is
same as the 1.5 rke profiles.
Tests that are marked manual but have an audit command that passes, have been changed to "automated"
Remaining fixes are to make some of the tests pass, same as the RKE 1.5 profiles.