Add WARN state, include unscored tests in report
This commit changes the report generation logic to include all tests, scored and unscored.
It also adds a new state "warn" corresponding to the "warn" output that kube-bench generates
for some of the manual tests.
It adds two new columns to the report, one to indicate if the test is scored/unscored and another
one to show if test is manual/automated