Changes for RKE 1.6 permissive profile
All tests that aren't applicable to RKE are "skipped" with the reason in remediation. This is
same as the 1.5 rke profiles.
Some tests that are applicable to RKE but fail on non-hardened clusters are "skipped". This is same as the
1.5 profile.
Tests that are marked manual but have an audit command that passes, have been changed to "automated"
Remaining fixes are to make some of the tests pass, same as the RKE 1.5 profiles.