build: Avoid version logic being executed twice The version logic is contained within hack/make/build.mk and is a pre-req for some of the Makefile targets.
The Makefile is used to kick off the building of the container image, and inside that
process make build is executed to compile security-scan. This change stops the version
logic from executing a second time by accepting a VERSION var to be set, therefore removing
the need for accessing the .git dir.

Signed-off-by: Paulo Gomes <paulo.gomes@suse.com>